Move common classes into api module
