+++ api/src/main/java/run/halo/app/core/extension/AnnotationSetting.java
