+++ api/src/main/java/run/halo/app/infra/Condition.java
