+++ api/src/main/java/run/halo/app/infra/ConditionStatus.java
