+++ application/build.gradle
@@ -43,6 +43,10 @@ bootJar {
     }
 }
 
+tasks.named('jar') {
+    enabled = false
+}
+
 ext {
     commonsLang3 = "3.12.0"
     base62 = "0.1.3"
@@ -57,55 +61,11 @@ ext {
 
 dependencies {
     implementation project(':api')
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-mail'
-    implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
-    implementation 'org.springframework.boot:spring-boot-starter-data-r2dbc'
-
-    // Spring Security
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.security:spring-security-oauth2-jose'
-    implementation 'org.springframework.security:spring-security-oauth2-client'
-    implementation 'org.springframework.security:spring-security-oauth2-resource-server'
-
-    implementation "org.springdoc:springdoc-openapi-starter-webflux-ui"
-    implementation 'org.openapi4j:openapi-schema-validator'
-    implementation "net.bytebuddy:byte-buddy"
-
-    // Apache Lucene
-    implementation "org.apache.lucene:lucene-core"
-    implementation "org.apache.lucene:lucene-queryparser"
-    implementation "org.apache.lucene:lucene-highlighter"
-    implementation "org.apache.lucene:lucene-backward-codecs"
-    implementation 'cn.shenyanchao.ik-analyzer:ik-analyzer'
-
-    implementation "org.apache.commons:commons-lang3"
-    implementation "io.seruco.encoding:base62"
-    implementation "org.pf4j:pf4j"
-    implementation "com.google.guava:guava"
-    implementation "org.jsoup:jsoup"
-    implementation "io.github.java-diff-utils:java-diff-utils"
-    implementation "org.springframework.integration:spring-integration-core"
-    implementation "com.github.java-json-tools:json-patch"
-    implementation "org.thymeleaf.extras:thymeleaf-extras-springsecurity6"
 
     annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
     annotationProcessor "org.springframework:spring-context-indexer"
 
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
-
-    // R2DBC
-    // Currently, official doesn't support mssql and mariadb yet until drivers are available.
-    // See https://github.com/spring-projects/spring-data-relational/commit/ee6c2c89b5c433748b22a79cf40dc8e01142caa3
-    // for more.
-    runtimeOnly 'io.r2dbc:r2dbc-h2'
-    runtimeOnly 'com.github.jasync-sql:jasync-r2dbc-mysql'
-    runtimeOnly 'org.mariadb:r2dbc-mariadb'
-    runtimeOnly 'org.postgresql:postgresql'
-    runtimeOnly 'org.postgresql:r2dbc-postgresql'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'org.springframework.security:spring-security-test'
